Change tracing interface
This removes the old tracing interface that relied on torch execution
hooks due to a lack of flexiblity. Torch hooks can only be set on an
`nn.Module` instance, which made the class hierarchy unnecessarily
complicated.

This change instead provides an internal tracing function that can be
triggered by populating the `traced_results` member of an `ApproxLayer`
instance. The tracing is then done within the `torch.autograd.Function`,
right before GeMM is called.